Revert "use standard types"
Breaks tmr tests.

This reverts commit c4f438616ac0b9b9a285be3f04ae4acef0cb0ae5.

+++ lib/include/types.h
@@ -1,13 +1,13 @@
 #ifndef  _TYPES_H
 #define _TYPES_H
 
-typedef signed char int8_t;
-typedef unsigned char uint8_t;
-typedef signed int int16_t;
-typedef unsigned int uint16_t;
-typedef signed long int int32_t;
-typedef unsigned long int uint32_t;
-typedef signed long long int int64_t;
-typedef unsigned long long int uint64_t; 
+typedef signed    char  int8_t;
+typedef unsigned  char  uint8_t;
+typedef signed    short int16_t;
+typedef unsigned  short uint16_t;
+typedef signed    long  int32_t;
+typedef unsigned  long  uint32_t;
+typedef signed    long  long int64_t;
+typedef unsigned  long  long uint64_t;
 
 #endif
